requirements.txt
pandas
openai
python-dotenv
boto3
